Refactor chat functionality into components
Extracted chat-related logic and UI elements into separate, reusable components to streamline the root index page. The refactoring includes creating distinct Chat, ChatLog, ChatDots, LoginForm, LogoutButton, PromptForm, and Layout components, encapsulating related concerns and improving readability and maintainability of code. This change also makes future updates and feature additions easier to manage by isolating chat functionality from the rest of the application.
