+++ app/src/components/Chat.tsx
@@ -1,0 +1,148 @@
+import { SupabaseClient } from "@supabase/supabase-js";
+import { useEffect, useRef, useState } from "react";
+import ChatLog, { Message } from "./ChatLog";
+import { toast } from "react-toastify";
+import LogoutButton from "./LogoutButton";
+import { Button } from "./ui/button";
+import { Plus } from "lucide-react";
+import { ThemeToggle } from "./ThemeToggle";
+import PromptForm from "./PromptForm";
+
+export default function Chat({
+  supabaseClient,
+}: {
+  supabaseClient: SupabaseClient;
+}) {
+  const bottomRef = useRef<HTMLDivElement | null>(null);
+  const textareaRef = useRef<HTMLTextAreaElement | null>(null);
+
+  const [entryData, setEntryData] = useState("");
+  const [messages, setMessages] = useState<Message[]>([]);
+  const [chatId, setChatId] = useState<Message[]>([]);
+  const [waitingForResponse, setWaitingForResponse] = useState(false);
+
+  const onSendMsgClick = async () => {
+    try {
+      let newMessages = [...messages, { role: "user", content: entryData }];
+      setMessages(newMessages);
+      setEntryData("");
+
+      setWaitingForResponse(true);
+      const { data: d2, error: e2 } = await supabaseClient
+        .from("conversations")
+        .update({ context: newMessages })
+        .eq("id", chatId);
+
+      if (e2) {
+        toast.error(e2.message || e2.code || "Unknown error");
+      }
+
+      const { data, error } = await supabaseClient.functions.invoke("chat", {
+        body: { messageHistory: newMessages },
+      });
+      setWaitingForResponse(false);
+      if (error) {
+        throw error;
+      }
+      setMessages([...newMessages, data?.msg]);
+    } catch (error: any) {
+      console.error("ERROR", error);
+      toast.error(error.message || error.code || error.msg || "Unknown error");
+    }
+  };
+
+  useEffect(() => {
+    if (messages.length > 1) {
+      bottomRef.current?.scrollIntoView({ behavior: "smooth", block: "end" });
+    }
+  }, [messages]);
+
+  const newConversation = async () => {
+    try {
+      const { data, error } = await supabaseClient
+        .from("conversations")
+        .insert([
+          {
+            context: [
+              { role: "assistant", content: "Hey, how can I help you?" },
+            ],
+          },
+        ])
+        .select("*");
+      if (error) {
+        console.error("ERROR", error);
+      }
+      if (!data || data.length == 0) {
+        throw new Error("No data returned");
+      }
+      console.log("data", data);
+      setMessages(data[0].context);
+      setChatId(data[0].id);
+    } catch (error: any) {
+      console.error("ERROR", error);
+      toast.error(error.message || error.code || error.msg || "Unknown error");
+    }
+  };
+
+  const fetchLastConversation = async () => {
+    try {
+      const { data, error } = await supabaseClient
+        .from("conversations")
+        .select("*")
+        .order("created_at", { ascending: false })
+        .limit(1);
+
+      if (!data || data.length == 0 || error) {
+        newConversation();
+      } else {
+        console.log("data", data);
+        setMessages(data[0].context);
+        setChatId(data[0].id);
+      }
+    } catch (error: any) {
+      console.error("ERROR", error);
+      toast.error(error.message || error.code || error.msg || "Unknown error");
+    }
+  };
+
+  useEffect(() => {
+    fetchLastConversation();
+  }, []);
+
+  return (
+    <>
+      <div className="fixed flex space-x-4 top-4 right-4">
+        <LogoutButton
+          supabaseClient={supabaseClient}
+        />
+        <Button
+          size={'icon'}
+          className="rounded-full bg-muted/20 text-muted-foreground hover:bg-muted/40"
+          onClick={async () => {
+            setMessages([]);
+            await newConversation();
+          }}
+        >
+          <Plus size={20} />
+        </Button>
+        <ThemeToggle />
+      </div>
+
+      <div className="p-8 mt-12 mb-32">
+        <ChatLog
+          messages={messages}
+          waitingForResponse={waitingForResponse}
+        />
+      </div>
+
+      <div ref={bottomRef} />
+      <PromptForm
+        textareaRef={textareaRef}
+        entryData={entryData}
+        setEntryData={setEntryData}
+        waitingForResponse={waitingForResponse}
+        onSendMsgClick={onSendMsgClick}
+      />
+    </>
+  );
+}
